requirements.txt
datasets==2.14.5
streamlit-audiorec==0.1.3
torch==2.2.1
accelerate==0.28.0
streamlit-jupyter==0.2.1
librosa==0.9.2
soundfile==0.12.1
gliner==0.1.4
pydantic==2.6.4
transformers==4.39.1